Added strobe to clear up latch-based regfile at startup
The latch-based register file is not sensitive to the rst_n and
requires an active clock to be cleared. Therefore, as soon as
a clock arrives at its interface, a 2-cycle clear strobe is
generated to set the regfile state at a known value.

This commit also fixes a problem where the write() and read()
methods of the hwpe_ctrl_intf_periph interface had the wrong
wen signal polarity.