Add addr mod, proper cuts, common_cells dep
Add an AXI4 address modification module which unpacks the address
signals in an AXI interface and allows them to be altered. This is
intended as a convenience module to allow for quick address changes.

Also add a proper AXI4 cut that interrupts all combinatorial paths. Drop
the spill register in favor of the one in common_cells. Add the latter
as a dependency.