Add clock domain crossing FIFO
Add the `cdc_fifo` built around the `cdc_2phase` primitive. The FIFO is
implemented in a way such that both the source and destination clock
domain maintain a read and write pointer. The source domain periodically
pushes its write pointer across the boundary, the destination domain
periodically pushes its read pointer across the boundary. Both do this
in a safe manner via the `cdc_2phase` primitive. Handshake signals at
the FIFO and towards the surrounding circuitry is generated purely based
on the pointers.