IPDatabase.py: don't checkout specified version if current version newer
This commit adds a new parameter `force_downgrade` and modifies
`update_ips()` to check the timestamp of the current IP version
and the version specified in the manifest. If the current IP version
is newer than the specified version, the checkout is not performed and
a warning is printed unless `force_downgrade` is set to true.

This is useful if experimenting with newer IP versions in hierarchical
flows such as PULPIssimo. Without this commit, the user has to first
update and commit & push `ips_list.yml` of the higher-level IP, and
then do the same with the top-level `ips_list.yml` which is a lot of
overhead. Otherwise, the tool refuses to work with a newer IP version
and always checks out the older version specified in the `ips_list.yml`
of the higher-level IP.