:bug: Fix bug in conflict resolution
Creating a list of removed_keys and then removing the indices from the list with
list.pop(i) does not work as the indices of the remaining entries in the list
change unless the last item in the list is removed. Instead remove the
duplicates by creating a list of objects and use the list.remove(obj) function.